Refactor package path for implementations
