+++ src/main/java/com/clutcher/comments/annotator/impl/JavaKeywordHighlighterAnnotator.java
@@ -1,5 +1,6 @@
-package com.clutcher.comments.annotator;
+package com.clutcher.comments.annotator.impl;
 
+import com.clutcher.comments.annotator.AbstractKeywordHighlighterAnnotator;
 import com.clutcher.comments.utils.AnnotatorUtils;
 import com.intellij.psi.PsiElement;
 import org.jetbrains.annotations.NotNull;
