+++ src/main/resources/META-INF/commentsHighlighter-withCSharp.xml
@@ -1,6 +1,6 @@
 <idea-plugin>
     <extensions defaultExtensionNs="com.intellij">
         <annotator language="C#"
-                   implementationClass="com.clutcher.comments.annotator.CSharpKeywordHighlighterAnnotator"/>
+                   implementationClass="com.clutcher.comments.annotator.impl.CSharpKeywordHighlighterAnnotator"/>
     </extensions>
 </idea-plugin>
