+++ src/main/resources/META-INF/commentsHighlighter-withPHP.xml
@@ -2,6 +2,6 @@
     <extensions defaultExtensionNs="com.intellij">
         <!-- ObjectiveC languageID is used for C and C++ also -->
         <annotator language="PHP"
-                   implementationClass="com.clutcher.comments.annotator.PHPKeywordHighlighterAnnotator"/>
+                   implementationClass="com.clutcher.comments.annotator.impl.PHPKeywordHighlighterAnnotator"/>
     </extensions>
 </idea-plugin>
